Tests/LibWeb/Text/expected/scroll-into-view-center.txt

The page has been scrolled to y: 800
